Refactor handling of repositories state from screen to useRepositories
